% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_lifestage.R
\name{consolidate_lifestage}
\alias{consolidate_lifestage}
\title{Consolidate LifeStage}
\usage{
consolidate_lifestage(df)
}
\arguments{
\item{df}{A dataframe of imported APPS data including the Species and LifeStage columns}
}
\description{
Consolidate detailed life stage categories into generalized juvenile and adult categories
}
\details{
Life stage categories of smolt, fry, yearling, sub-yearling and parr are recoded as "juveniles" for summary analyses. Jacks and subadult salmon and steelhead are recoded as "adults", though their reproductive value to the species is expected to be less than fully mature adults. Kelt steelhead are not recoded as adults, but left separate. The egg lifestage is removed for salmon and steelhead but retained for sturgeon, rockfish, and eulachon, as is the subadult category.
}
