% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_speciesvar.R
\name{create_speciesvar}
\alias{create_speciesvar}
\title{Create species variable}
\usage{
create_speciesvar(df)
}
\description{
Modify species common names to reader-friendly format and combine with population name to get ESU or DPS unit name
}
\details{
This function first re-orders the CommonName field exported from APPS, then concatenates it with the Population
}
