% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_totalactmorts.R
\name{create_totalactmorts}
\alias{create_totalactmorts}
\title{Create Total Actual Morts Column}
\usage{
create_totalactmorts(df)
}
\description{
Create a new column summing the total actual/reported mortalities for each take line
}
\details{
The total number of actual mortalities is the sum of reported uninentional lethal takes plus the sum of reported intentional lethal takes. This function identifies whether the Take Action is Intentional/Directed Mortality, and counts reported takes as mortalities in those cases.
}
