% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-lsOrder.R
\docType{data}
\name{ls.order}
\alias{ls.order}
\title{LifeStage Order data object}
\format{
ordered list
}
\usage{
data(ls.order)
}
\description{
Order we want life stages to appear in formatted tables, currently listed oldest to youngest
}
\keyword{datasets}
