% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_table.R
\name{order_table}
\alias{order_table}
\title{Order Table}
\usage{
order_table(df)
}
\arguments{
\item{df}{A dataframe of imported APPS data that you want to put in print order, including the Species, LifeStage, and Production columns}
}
\description{
Order a dataframe or table by Species, LifeStage, and Production
}
\details{
This arrangement is set in a North to South species pattern. Currently set up to create ordered factors out of these columns first, then order all at once.
}
