% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prOrder.R
\docType{data}
\name{pr.order}
\alias{pr.order}
\title{Production Order data object}
\format{
ordered list
}
\usage{
data(pr.order)
}
\description{
Order we want origin to appear in formatted tables, currently listing natural-origin first
}
\keyword{datasets}
