% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-spOrder.R
\docType{data}
\name{sp.order}
\alias{sp.order}
\title{Species Order data object}
\format{
ordered list
}
\usage{
data(sp.order)
}
\description{
Order we want species to appear in formatted tables, currently listed North to South
}
\keyword{datasets}
