% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbrev_takeaction.R
\name{abbrev_takeaction}
\alias{abbrev_takeaction}
\title{Abbreviate Take Action}
\usage{
abbrev_takeaction(df)
}
\arguments{
\item{df}{A dataframe of imported APPS data including the Species and TakeAction columns}
}
\description{
Abbreviate take actions to reader-friendly format for tables
}
\details{
Used only for formatting values for print-friendly tables, but shouldn't be implemented during data management
}
