% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perc_ESU_DPS.R
\name{perc_ESU_DPS}
\alias{perc_ESU_DPS}
\title{Percent of ESU/DPS}
\usage{
perc_ESU_DPS(df, abund)
}
\description{
Calculate the percent of the ESU or DPS abundance that will be taken or killed by proposed research
}
\details{
A dataframe of proposed take is merged with a dataframe of abundance to calculate the percentage of abundance the take summary represents. This can be applied to an individual permit summary or a baseline or program-wide take summary as long as the necessary columns (Species, LifeStage, Production, ExpTake, TotalMorts, abundance) are present
}
