% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_flextable.R
\name{pretty_flextable}
\alias{pretty_flextable}
\title{Pretty flextable}
\usage{
pretty_flextable(df, ftname)
}
\description{
Takes a tibble, creates pretty, formatted flextable for export into word docs
}
\details{
This function only FORMATS the flextable - you must open a docx, add flextables to the body, and then PRINT the docx file to export, although a preview can be printed if desired
}
