% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_commonname.R
\name{rename_commonname}
\alias{rename_commonname}
\title{Rename CommonName}
\usage{
rename_commonname(df)
}
\arguments{
\item{df}{A dataframe of imported APPS data with an unaltered CommonName column that you want to re-format into print-friendly order}
}
\description{
Modify species common names to reader-friendly format and combine with population name to get ESU or DPS unit name
}
\details{
takes a dataframe that already has the CommonName variable and updates the wording FROM the left hand side of the ~ TO the right hand side of the ~
}
